[WIP] Enable dotty and scala dependencies on the same setup
- unifiying scalac and dotc benchmarks (relied on @jvican's port!)
