+++ compilation/src/main/scala/base/JavacBenchmark.java
@@ -1,4 +1,4 @@
-package scala.tools.nsc;
+package base;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Scope;
