+++ compilation/src/main/scala/dotty/tools/dotc/WarmScalacBenchmark.scala
@@ -1,4 +1,4 @@
-package scala.tools.nsc
+package dotty.tools.dotc
 
 import java.util.concurrent.TimeUnit
 
@@ -11,10 +11,10 @@ import org.openjdk.jmh.annotations._
   */
 @BenchmarkMode(Array(SampleTime))
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
-@Warmup(iterations = 8, time = 10, timeUnit = TimeUnit.SECONDS)
-@Measurement(iterations = 10, time = 10, timeUnit = TimeUnit.SECONDS)
+@Warmup(iterations = 0)
+@Measurement(iterations = 1, time = 30, timeUnit = TimeUnit.SECONDS)
 @Fork(value = 3)
-class HotScalacBenchmark extends BaseBenchmark {
+class WarmDotcBenchmark extends BaseBenchmark {
   @Benchmark
-  override def compileScalac(): Unit = super.compileScalac()
+  override def compileDotc(): Unit = super.compileDotc()
 }
